Properly replace line on template application
When a template was applied, we were simply setting the 'content'
property of the existing first block of the canvas that wasn't the
title. This was causing problems because we would end up with a type
mismatch in the UI. Now, we fully replace the block and then insert as
normal after that.

- Fixes usecanvas/pro-web#120
- Fixes usecanvas/pro-web#126